Huge update. stubs of methods, some implementations, new constants... read the description.
Added more constants ( types for the "datatype" field)

Added "DataPacket" object to utilities project

Added stubs of methods with commented plain english strategies for the
following methods in Utils:
- sendFileTo:: give a udpclient as a target and a filename, sends the
file to that target.
- getChecksum:: give a byte array, get that array's checksum
- getfileMetadataPacket:: give filename, type of metadata packet,
receive a full byte array for whatever you need to use it for. Asking
the server for a file, call this with the type as the constant
TYPE_GET_REQUEST, along with the name of the file. the packet returned
has those values and the length of that name, leaving the filesize etc.
fields as 0's. If the type is, say, TYPE_INCOMING, then you are creating
a packet so you can send a file. This method will fill out the filesize,
packet total, and everything else so the reciever will know what file it
is going to recieve. TYPE_DIRECTORY_REQUEST and TYPE_DIRECTORY_DELIVERY
are not yet implemented.

Possibly wrote a bunch of good stuff. Also maybe just wrote a bunch of
junk. I dunno.